Parse input and output in project scripts

Parses project diretory to identify input andoutput variables, basenames, etc.

parse_proj_io(paths, dir, ..., rgx_file_io = ".R",
  rgx_input = ".*(import_ext|read)", rgx_output = ".*export_ext|write)")

Arguments

paths
character. Can be a vector. Missing by default. Takes precedence over dir argument.

dir
character. Should not be a vector. Missing by default. Only used if paths is missing.

...
dots. Arguments passed to list.files() for identifying input/output files. (Technically, pattern could/should not be specified explicitly.)

rgx_file_io
character. Alias to pattern parameter for list.files(). Used ONLY if paths is missing and dir is not.

rgx_input
character. Regular expression to match for input functions

rgx_output
character. Regular expression to match for output functions.

Value

data.frame.

Details

Intended to be used with this package's export() and import() functions, but it can also work with the functions from other packages (e.g. readr) if rgx_input and rgx_output are modified appropriately. For example, rgx_input == "read" and rgx_output == "write" might be appropriate choices if using the readr package.) Does not (currently) work as well if variables, basenames, etc. are not wrapped directly by the functions described by rgx_input and rgx_output. NOTE: In the future this should/will use utils::getParseData(parse(f))