Render project output

Renders R scripts with Roxygen comments into specified output format.

render_proj_io(path_input, dir_input, ..., dir_output = file.path(getwd()),
  basename_output, rgx_input = ".", rgx_input_include = rgx_input,
  rgx_input_exclude = "", rgx_output_trim = "-v[0-9]+.R",
  ext_output = c("html"), render = TRUE, render_params = NULL,
  return = TRUE, overwrite = TRUE, quiet = FALSE, backup = FALSE,
  backup_suffix = paste0("_", format(Sys.time(), "%Y-%m-%d_%H-%M-%S")),
  keep_rmd = FALSE)

Arguments

path_input
character. Can be a vector.

dir_input
character. Should not be a vector. Only used if path_input is missing.

...
dots. Parameters passed to list.files().

dir_output
character. Should not be a vector. IMPORTANT: Relative to input directory (so something like "../output" is valid).

basename_output
character. Explicit basenames to use for output. Length of variable must exactly match number of path meeting list.files() criteria.

rgx_input
character. Alias to pattern parameter for list.files(). Used ONLY if path is missing and dir is not.

rgx_input_include
character. Regular expression to use to filter for list.files() output. (Somewhat redundant.)

rgx_input_exclude
character. Regular expression to use to filter for list.files() output. (Somewhat redundant.)

rgx_output_trim
character. Used ONLY if basename_output is missing. Describes how input file names should be "trimmed" to make the output file name appear "cleaner".

ext_output
character. File extension of output.

render
logical. Indiciates whether or not to actually carry ret function.

render_params
list. Parameters to pass directly to params argument of rmarkdown::render()

return
logical. Relevant ONLY if render == FALSE. Set to TRUE in order to preview what would be rendered.

overwrite
logical. Indicates whether or not to overwrite any existing file with the same name. Not currently used.

quiet
logical. Direct argument for rmarkdown::render().

backup
logical. Indicates whether or not to create e a backup.

backup_suffix
character. Suffix to append to file for backup file.

keep_rmd
logical. Assuming specified output is not ".Rmd", indicates whether to call knitr::spin to keep "intermediate" results.

Value

data.frame. Information regarding what was rendered.

Details

This is a wrapper for rmarkdown::render()` (and knitr::spin(), if keep_rmd == TRUE). It is designed specifically to convert R scripts formatted with Roxgyen comments into the specified output format. (In other words, it mirrors the behavior of knitr::knit() for a R markdown document.